services/auth/requirements.txt
# Auth service dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
redis[asyncio]>=5.0.0
resend>=0.8.0
